requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-groq>=0.1.0
pydantic>=2.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-dotenv>=1.0.0
protobuf==3.20.3
livekit-api>=0.8.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-pptx
Pillow
python-jose[cryptography]
bcrypt
pytest
httpx
